moved vision provider to autonomous
